feat(6828): extract data layer postgres execution adapter tests
